driver: remember about pending tx kicks
Because of skb->xmit_more, it may happen that txp asks for a kick,
but the kick is not sent because skb->xmit_more is true. When
finally a packed with skb->xmit_more set to false, txp may not
ask for a kick. The end result is a lost notification. This bug
shows up all the time with event idx.